+++ CHANGELOG.md
@@ -61,6 +61,7 @@ NetBox now supports modeling physical cables for console, power, and interface c
 * [#2601](https://github.com/digitalocean/netbox/issues/2601) - Added a `description` field to pass-through ports
 * [#2602](https://github.com/digitalocean/netbox/issues/2602) - Return HTTP 204 when no new IPs/prefixes are available for provisioning
 * [#2608](https://github.com/digitalocean/netbox/issues/2608) - Fixed null `outer_unit` error on rack import
+* [#2609](https://github.com/digitalocean/netbox/issues/2609) - Fixed exception when ChoiceField integer value is passed as a string
 
 ## API Changes